Moved evaluate function
